+++ src/plone/jsonapi/core/browser/api.py
@@ -10,7 +10,9 @@ from Products.Five import BrowserView
 from zope.publisher.interfaces import IPublishTraverse
 
 from decorators import runtime
+from decorators import returns_xml
 from decorators import returns_json
+from decorators import returns_binary_stream
 from decorators import handle_errors
 
 from interfaces import IAPI
@@ -55,10 +57,24 @@ class API(BrowserView):
     @returns_json
     @runtime
     @handle_errors
-    def render(self):
+    def to_json(self):
+        return self.dispatch()
+
+    @returns_binary_stream
+    def to_binary_stream(self):
+        return self.dispatch()
+
+    @returns_xml
+    def to_xml(self):
         return self.dispatch()
 
     def __call__(self):
         """ render json on __call__
         """
-        return self.render()
+        accept = self.request.getHeader("Accept")
+        if self.request.form.get("asbinary", False) or accept == "application/zip":
+            return self.to_binary_stream()
+        if self.request.form.get("asxml", False) or accept == "application/xml":
+            return self.to_xml()
+        # return JSON per default
+        return self.to_json()